Fix workers being broken due to trying to import electron
